+++ src/_/css.ts
@@ -1,21 +1,9 @@
-declare const _RawCssString__Brand: unique symbol;
+declare const _CssString__Brand: unique symbol;
 /**
  * Raw, un-transpiled CSS code.
  * May be incomplete and/or include (SCSS-like) nested syntax
- *
- * Soft-branded as a raw CSS string, only to convey intent/expectation —
- * not to prevent other strings being passed.
  */
-export type RawCssString = string & { [_RawCssString__Brand]?: true };
-
-declare const _RawCssValue__Brand: unique symbol;
-/**
- * Raw, rendered (string-cast) CSS value.
- *
- * Soft-branded as a raw CSS value, only to convey intent/expectation —
- * not to prevent other strings being passed.
- */
-export type RawCssValue = string & { [_RawCssValue__Brand]?: true };
+export type CssString = string & { [_CssString__Brand]: true };
 
 /** Filters out all falsy values except `0`. */
 const notFalsy = (val: unknown) => !!val || val === 0;
@@ -32,7 +20,7 @@ const notFalsy = (val: unknown) => !!val || val === 0;
 export const css = function (
   strings: TemplateStringsArray,
   ...values: Array<unknown>
-): RawCssString {
+): CssString {
   const valCount = values.length;
   return strings
     .map((str, i) => {
@@ -48,13 +36,13 @@ export const css = function (
 
       return str + String(notFalsy(value) ? value : '');
     })
-    .join('');
+    .join('') as CssString;
 };
 
 // ---------------------------------------------------------------------------
 
 /**
- * An alias for the ` css`` ` templater, for cases where you're writing a
+ * An alias for the ` css`` ` templater function, for cases where you're writing a
  * standalone CSS value or other out-of-context CSS snippets, and you wish to
  * disable VSCode's syntax highlighting and error checking.
  *
@@ -64,12 +52,13 @@ export const cssVal = css;
 
 // ---------------------------------------------------------------------------
 
-declare const _RawMediaQuery__Brand: unique symbol;
+declare const _CssediaQueryString__Brand: unique symbol;
 /**
- * String, but soft-branded as a media query.
- * Only used to convey intent/expectation — not to prevent other strings being passed.
+ * A CSS media-query string.
+ *
+ * (Example: `"screen and (max-width: 760px)""`
  */
-export type RawMediaQuery = string & { [_RawMediaQuery__Brand]?: true };
+export type CssMediaQueryString = string & { [_CssediaQueryString__Brand]: true };
 
 /**
  * @deprecated  Use DIY `@at-root` wraps instead  (Will be removed in v0.6+)
@@ -80,11 +69,19 @@ export type RawMediaQuery = string & { [_RawMediaQuery__Brand]?: true };
  * This mixin is really only useful as a defence mechanism inside CSS-emitting
  * functions (mixins) that will be used in unpredictable `@media` contexts.
  */
-export function media(query: RawMediaQuery): (cssContent: RawCssString) => RawCssString;
-export function media(query: RawMediaQuery, cssContent: RawCssString): RawCssString;
+export function media(
+  query: CssMediaQueryString | string
+): (cssContent: CssString | string) => CssString;
+export function media(
+  query: CssMediaQueryString | string,
+  cssContent: CssString | string
+): CssString;
 
-export function media(query: RawMediaQuery, cssContent?: RawCssString) {
-  const mediaWrapper = (cssContent: RawCssString) =>
+export function media(
+  query: CssMediaQueryString | string,
+  cssContent?: CssString | string
+) {
+  const mediaWrapper = (cssContent: CssString | string) =>
     css`
       ${'@at-root (without: media)'} {
         @media ${query} {
@@ -107,3 +104,14 @@ export function media(query: RawMediaQuery, cssContent?: RawCssString) {
  * @see https://github.com/maranomynet/es-in-css/tree/v0.5#str-quoted-string-printer
  */
 export const str = (string: string) => JSON.stringify(`${string}`);
+
+// ===========================================================================
+// Deprecated types:
+
+declare const _RawCssString__Brand: unique symbol;
+/** @deprecated  Use `CssString` type instead (Will be removed in v0.8) */
+export type RawCssString = string & { [_RawCssString__Brand]?: true };
+/** @deprecated  Use either `RawCssString` or just `string` instead (Will be removed in v0.8) */
+export type RawCssValue = string & { [_RawCssString__Brand]?: true };
+/** @deprecated  Use `CssMediaQueryString` type instead (Will be removed in v0.8) */
+export type RawMediaQuery = string & { [_RawCssString__Brand]: true };