+++ src/_/index.tests.ts
@@ -65,8 +65,10 @@ o.spec('es-in-css entry point', () => {
 /* eslint-disable @typescript-eslint/no-unused-vars, unused-imports/no-unused-imports-ts, import/first, simple-import-sort/imports */
 import type {
   // css.ts
+  CssString,
+  CssVarString,
+  CssMediaQueryString,
   RawCssString,
-  RawCssValue,
   RawCssVarString,
   RawMediaQuery,